requirements.txt
astroid==2.5.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==7.1.2
Cython==0.29.33
Flask==1.1.2
idna==3.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.23
numpy==1.24.2
pandas==1.5.3
protobuf==3.15.5
pylint==2.7.2
python-dateutil==2.8.2
python-math==0.0.1
pytz==2023.3
requests==2.28.2
six==1.15.0
toml==0.10.2
urllib3==1.26.15
Werkzeug==1.0.1
wrapt==1.12.1
